update: MainAppをStatefulWidgetに変更
